Lint error fixes
